requirements.txt
sentence-transformers==2.7.0
scikit-learn==1.5.0
nltk==3.8.1
pyMuPDF==1.24.9
pdfplumber==0.1.0
pandas==2.0.3
numpy==1.26.4
scipy==1.13.1
tqdm==4.66.4
torch==2.3.1
transformers==4.41.2
rapidfuzz
fastapi==0.111.0
uvicorn==0.29.0
firebase-admin==6.5.0
streamlit==1.35.0
requests==2.32.3
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.7.1
httpx==0.27.0
PyJWT==2.8.0
pycryptodome==3.20.0
bcrypt==4.1.2
python-jose==3.5.0